+++ src/command.rs
@@ -94,6 +94,18 @@ mod main;
 ///         }
 ///         Ok(())
 ///     }
+///
+///     fn into_any(self: Box<Self>) -> Box<dyn std::any::Any> {
+///         self
+///     }
+///
+///     fn as_any(&self) -> &dyn std::any::Any {
+///         self
+///     }
+///
+///     fn as_any_mut(&mut self) -> &mut dyn std::any::Any {
+///         self
+///     }
 /// }
 ///
 /// fn main() -> Result<()> {