+++ src/subcommand/dist.rs
@@ -1,3 +1,5 @@
+use std::any::Any;
+
 use crate::{config::Config, Result, Run};
 
 /// Arguments definition of the `dist` subcommand.
@@ -19,12 +21,24 @@ impl Run for Dist {
     fn run(&self, config: &Config) -> Result<()> {
         self.run(config)
     }
+
+    fn into_any(self: Box<Self>) -> Box<dyn Any> {
+        self
+    }
+
+    fn as_any(&self) -> &dyn Any {
+        self
+    }
+
+    fn as_any_mut(&mut self) -> &mut dyn Any {
+        self
+    }
 }
 
 impl Dist {
     /// Runs the `dist` subcommand.
     #[tracing::instrument(name = "dist", skip_all, err)]
-    pub(crate) fn run(&self, config: &Config) -> Result<()> {
+    pub fn run(&self, config: &Config) -> Result<()> {
         let Self {
             #[cfg(subcommand_dist_build)]
             dist_build_args,