+++ src/subcommand/fmt.rs
@@ -1,4 +1,4 @@
-use std::process::Command;
+use std::{any::Any, process::Command};
 
 use crate::{
     args::{EnvArgs, PackageArgs},
@@ -25,6 +25,18 @@ pub struct Fmt {
 impl Run for Fmt {
     fn run(&self, config: &Config) -> Result<()> {
         self.run(config)
+    }
+
+    fn into_any(self: Box<Self>) -> Box<dyn Any> {
+        self
+    }
+
+    fn as_any(&self) -> &dyn Any {
+        self
+    }
+
+    fn as_any_mut(&mut self) -> &mut dyn Any {
+        self
     }
 }